Correction of many writing errors
Hey. It's Liam Tailor. I've edited a lot of punctuation, ortography, esthethic and many other writing errors. I have also done a few slight changes, to appeal better to the reader. Currently, all of the introduction to Cities at Night is very scientific and it lacks the appeal to a regular Citizen, especially the emotional impact. I'll try to further change that in later edits. I'm going to concentrate on this, rather than creating a polish tutorial, of wich I wrote in an e-mail, as it's a much more important task to make the site speak out to the general public better. One more thing. I've noticed some spelling errors on the main revolving bar on the top of citiesatnight.org, like "astronaut may be taken a picture" instead of "taking the picture". Wich file contains this, where can I correct that?
Contact me at liado@wp.pl